Add SWANpost GUI tool for folder and POI postprocessing
